refactor(myinfo-types.ts): move into types folder
